+++ tests/test_control.py
@@ -1,14 +1,14 @@
 # -*- coding: utf-8 -*-
-from datetime import datetime, timezone
 import platform
 import time
 import types
+from datetime import datetime, timezone
 from unittest.mock import MagicMock, patch
 
-from chaoslib.notification import RunFlowEvent
 import pytest
 import requests
 import requests_mock
+from chaoslib.notification import RunFlowEvent
 
 from chaoshumio import push_to_humio
 from chaoshumio.control import configure_control, with_logging
@@ -39,7 +39,7 @@ def test_push_context_to_humio():
     assert with_logging.enabled is True
     with requests_mock.mock() as m:
         m.post(
-            'https://cloud.humio.com/api/v1/dataspaces/default/ingest',
+            "https://cloud.humio.com/api/v1/ingest/humio-ingest",
             status_code=200,
             json=[ {
                 "events": [
@@ -51,6 +51,63 @@ def test_push_context_to_humio():
             } ]
         )
         push_to_humio(event, secrets["humio"])
+
+        assert m.called
+        payload = m.last_request.json()[0]
+        assert "tags" in payload
+        assert payload["tags"] == {
+            "chaosengineering": "true",
+            'host': node, 'level': 'an-event',
+            "platform": platform.platform(),
+            "python": platform.python_version(),
+            "system": platform.system(),
+            "machine": platform.machine(),
+            'provider': 'chaostoolkit', 'type': 'experiment'
+        }
+        assert payload["events"][0]["attributes"] == context
+
+
+def test_push_context_to_humio_custom_domain():
+    timestamp = time.time()
+    isotimestamp = datetime.fromtimestamp(timestamp, timezone.utc).isoformat()
+
+    node = platform.node()
+    context = {
+        "k0": "v0"
+    }
+    event = {
+        "ts": timestamp,
+        "name": "an-event",
+        "type": "experiment",
+        "context": context
+    }
+    secrets = {
+        "humio": {
+            "token": "1234",
+            "dataspace": "default"
+        }
+    }
+    humio_test_url = "https://my.humio.com"
+    configuration = {
+        "humio_url": humio_test_url
+    }
+
+    configure_control(None, secrets)
+    assert with_logging.enabled is True
+    with requests_mock.mock() as m:
+        m.post(
+            "{}/api/v1/ingest/humio-ingest".format(humio_test_url),
+            status_code=200,
+            json=[ {
+                "events": [
+                    {
+                        "timestamp": isotimestamp,
+                        "attributes": context
+                    },
+                ]
+            } ]
+        )
+        push_to_humio(event, secrets["humio"], configuration)
 
         assert m.called
         payload = m.last_request.json()[0]
@@ -95,7 +152,7 @@ def test_push_state_to_humio():
     assert with_logging.enabled is True
     with requests_mock.mock() as m:
         m.post(
-            'https://cloud.humio.com/api/v1/dataspaces/default/ingest',
+            'https://cloud.humio.com/api/v1/ingest/humio-ingest',
             status_code=200,
             json=[ {
                 "events": [
@@ -146,11 +203,12 @@ def test_logger_disabled_when_missing_token():
             "dataspace": "default"
         }
     }
+
     configure_control(None, secrets)
     assert with_logging.enabled is False
     with requests_mock.mock() as m:
         m.post(
-            'https://cloud.humio.com/api/v1/dataspaces/default/ingest',
+            'https://cloud.humio.com/api/v1/ingest/humio-ingest',
             status_code=200,
             json=[ {
                 "events": [
@@ -193,7 +251,7 @@ def test_logger_disabled_when_missing_dataspace():
     assert with_logging.enabled is False
     with requests_mock.mock() as m:
         m.post(
-            'https://cloud.humio.com/api/v1/dataspaces/default/ingest',
+            'https://cloud.humio.com/api/v1/ingest/humio-ingest',
             status_code=200,
             json=[ {
                 "events": [
@@ -205,4 +263,4 @@ def test_logger_disabled_when_missing_dataspace():
             } ]
         )
         push_to_humio(event, secrets["humio"])
-        assert m.called is False
+        assert m.called is True