+++ tests/test_notification.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
-from datetime import datetime, timezone
 import time
 import types
+from datetime import datetime, timezone
 from unittest.mock import MagicMock, patch
 
-from chaoslib.notification import RunFlowEvent
 import pytest
 import requests
 import requests_mock
+from chaoslib.notification import RunFlowEvent
 
 from chaoshumio.notification import notify
 
@@ -27,7 +27,50 @@ def test_notify():
     }
     with requests_mock.mock() as m:
         m.post(
-            'https://cloud.humio.com/api/v1/dataspaces/my-space/ingest',
+            'https://cloud.humio.com/api/v1/ingest/humio-ingest',
+            status_code=200,
+            json=[ { 
+                "tags": {
+                    "host": "fake-host"
+                },
+                "events": [
+                    {
+                        "timestamp": isotimestamp,
+                        "attributes": payload
+                    },
+                ]
+            } ]
+        )
+
+        notify(
+            {
+                "token": "my-token"
+            },
+            event_payload
+        )
+
+        assert m.called
+
+
+def test_notify_custom_URL():
+    payload = {
+        "msg": "hello"
+    }
+    timestamp = time.time()
+    isotimestamp = datetime.fromtimestamp(timestamp, timezone.utc).isoformat()
+
+    event_payload = {
+        "ts": timestamp,
+        "event": str(RunFlowEvent.RunStarted),
+        "phase": "run",
+        "payload": payload
+    }
+
+    humio_url = "https://myhumio.company.com"
+
+    with requests_mock.mock() as m:
+        m.post(
+            "{}/api/v1/ingest/humio-ingest".format(humio_url),
             status_code=200,
             json=[ { 
                 "tags": {
@@ -45,7 +88,7 @@ def test_notify():
         notify(
             {
                 "token": "my-token",
-                "dataspace": "my-space"
+                "humio_url": humio_url
             },
             event_payload
         )